+++ rust/Cargo.toml
@@ -56,6 +56,8 @@ arrow-schema = "47.0"
 arrow-select = "47.0"
 async-recursion = "1.0"
 async-trait = "0.1"
+aws-config = "0.56"
+aws-credential-types = "0.56"
 aws-sdk-dynamodb = "0.33"
 half = { "version" = "=2.3.1", default-features = false, features = [
     "num-traits",
