+++ rust/lance-core/Cargo.toml
@@ -22,6 +22,8 @@ arrow-schema.workspace = true
 arrow-select.workspace = true
 async-recursion.workspace = true
 async-trait.workspace = true
+aws-config.workspace = true
+aws-credential-types.workspace = true
 byteorder.workspace = true
 bytes.workspace = true
 chrono.workspace = true
