+++ rust/lance/src/io/commit.rs
@@ -53,6 +53,9 @@ use crate::dataset::{write_manifest_file, ManifestWriteConfig};
 use crate::format::{DeletionFile, Fragment};
 use crate::Dataset;
 
+#[cfg(all(target_feature="dynamodb", tests))]
+mod dynamodb;
+#[cfg(tests)]
 mod external_manifest;
 pub use lance_core::io::commit::latest_manifest_path;
